requirements.txt
dateutils
geojson
pytest
pytest-cov
pytest-pep8
pytest-pythonpath
rethinkdb
sphinx
sphinx_rtd_theme
tox
